backend/requirements.txt
django
requests
python-dotenv
django-cors-headers
gunicorn